Notifications now work for configuration,pugins,defintions
